Desktop/motivate/main/requirements.txt
amqp==5.1.0
appdirs==1.4.4
arabic-reshaper==2.1.3
arrow==1.2.1
asgiref==3.5.0
asn1crypto==1.5.1
atomicwrites==1.4.0
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
billiard==3.6.4.0
black==20.8b1
boto3==1.20.37
botocore==1.23.54
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.1.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
codacy-coverage==1.3.11
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cryptography==36.0.1
cssselect2==0.5.0
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.11
django-cors-headers==3.11.0
django-filter==21.1
django-mathfilters==1.0.0
django-phonenumber-field==6.0.0
django-ses==2.3.1
django-settings-export==1.2.1
django-storages==1.12.3
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-compound-fields==2.0.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
extension==1.0.0
Flask==2.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.1
jmespath==0.10.0
kombu==5.2.4
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
MouseInfo==0.1.3
mypy-extensions==0.4.3
openpyxl==3.0.9
oscrypto==1.3.0
packaging==21.3
pathspec==0.9.0
pdfkit==1.0.0
phonenumbers==8.12.41
Pillow==9.0.1
pip-check==2.7
pipdeptree==2.2.1
pluggy==1.0.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
PyJWT
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.7
PyPDF3==1.0.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytest==7.1.1
pytest-django==4.5.2
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytoml==0.1.21
pytweening==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.1.1
regex==2022.3.15
reportlab==3.6.9
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
sentry-sdk==1.5.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
terminaltables==3.1.10
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
twilio==7.8.1
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typed-ast==1.5.2
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.0
whitenoise==5.3.0
wrapt==1.14.0
xhtml2pdf==0.2.7
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
zope.interface==5.4.0
